Abstracted all message handlers in the worker
