+++ lib/wasmboy/worker/wasmboy.worker.js
@@ -7,6 +7,14 @@ import { postMessage, onMessage } from '../../worker/workerapi';
 import { WORKER_MESSAGE_TYPE, WORKER_ID } from '../../worker/constants';
 import { getEventData } from '../../worker/util';
 import { getSmartWorkerMessage } from '../../worker/smartworker';
+
+// Post message handlers
+import { graphicsWorkerOnMessage } from './graphics/onmessage';
+import { audioWorkerOnMessage } from './audio/onmessage';
+import { controllerWorkerOnMessage } from './controller/onmessage';
+import { memoryWorkerOnMessage } from './memory/onmessage';
+
+// Working the emalator
 import { instantiateWasm } from './instantiate';
 import { update } from './update';
 
@@ -61,16 +69,16 @@ libWorker = {
         // Assign our worker ports on connect
         if (eventData.message.workerId === WORKER_ID.GRAPHICS) {
           libWorker.graphicsWorkerPort = event.ports[0];
-          onMessage(libWorker._graphicsMessageHandler, libWorker.graphicsWorkerPort);
+          onMessage(graphicsWorkerOnMessage.bind(this, libWorker), libWorker.graphicsWorkerPort);
         } else if (eventData.message.workerId === WORKER_ID.MEMORY) {
           libWorker.memoryWorkerPort = event.ports[0];
-          onMessage(libWorker._memoryMessageHandler, libWorker.memoryWorkerPort);
+          onMessage(memoryWorkerOnMessage.bind(this, libWorker), libWorker.memoryWorkerPort);
         } else if (eventData.message.workerId === WORKER_ID.CONTROLLER) {
           libWorker.controllerWorkerPort = event.ports[0];
-          onMessage(libWorker._controllerMessageHandler, libWorker.controllerWorkerPort);
+          onMessage(controllerWorkerOnMessage.bind(this, libWorker), libWorker.controllerWorkerPort);
         } else if (eventData.message.workerId === WORKER_ID.AUDIO) {
           libWorker.audioWorkerPort = event.ports[0];
-          onMessage(libWorker._audioMessageHandler, libWorker.audioWorkerPort);
+          onMessage(audioWorkerOnMessage.bind(this, libWorker), libWorker.audioWorkerPort);
         }
 
         // Simply post back that we are ready
@@ -157,140 +165,6 @@ libWorker = {
     }
 
     return 0;
-  },
-
-  _graphicsMessageHandler: event => {
-    // Handle our messages from the main thread
-    const eventData = getEventData(event);
-
-    switch (eventData.message.type) {
-      case WORKER_MESSAGE_TYPE.GET_CONSTANTS: {
-        libWorker.WASMBOY_CURRENT_FRAME_OUTPUT_LOCATION = libWorker.wasmInstance.exports.frameInProgressVideoOutputLocation.valueOf();
-        libWorker.WASMBOY_CURRENT_FRAME_SIZE = libWorker.wasmInstance.exports.FRAME_SIZE.valueOf();
-        // Forward to our lib worker
-        libWorker.graphicsWorkerPort.postMessage(
-          getSmartWorkerMessage(
-            {
-              type: WORKER_MESSAGE_TYPE.GET_CONSTANTS_DONE,
-              WASMBOY_CURRENT_FRAME_OUTPUT_LOCATION: libWorker.wasmInstance.exports.frameInProgressVideoOutputLocation.valueOf()
-            },
-            eventData.messageId
-          )
-        );
-        return;
-      }
-    }
-  },
-
-  _memoryMessageHandler: event => {
-    // Handle our messages from the main thread
-    const eventData = getEventData(event);
-
-    switch (eventData.message.type) {
-      case WORKER_MESSAGE_TYPE.CLEAR_MEMORY: {
-        // Clear Wasm memory
-        // https://docs.google.com/spreadsheets/d/17xrEzJk5-sCB9J2mMJcVnzhbE-XH_NvczVSQH9OHvRk/edit?usp=sharing
-        for (let i = 0; i <= libWorker.wasmByteMemory.length; i++) {
-          libWorker.wasmByteMemory[i] = 0;
-        }
-
-        const wasmByteMemory = libWorker.wasmByteMemory.slice(0);
-        libWorker.memoryWorkerPort.postMessage(
-          getSmartWorkerMessage(
-            {
-              type: WORKER_MESSAGE_TYPE.CLEAR_MEMORY_DONE,
-              wasmByteMemory: wasmByteMemory.buffer
-            },
-            eventData.messageId
-          ),
-          [wasmByteMemory.buffer]
-        );
-        return;
-      }
-
-      case WORKER_MESSAGE_TYPE.GET_CONSTANTS: {
-        libWorker.WASMBOY_GAME_BYTES_LOCATION = libWorker.wasmInstance.exports.gameBytesLocation.valueOf();
-
-        // Forward to our lib worker
-        libWorker.memoryWorkerPort.postMessage(
-          getSmartWorkerMessage(
-            {
-              type: WORKER_MESSAGE_TYPE.GET_CONSTANTS_DONE,
-              WASMBOY_GAME_BYTES_LOCATION: libWorker.wasmInstance.exports.gameBytesLocation.valueOf(),
-              WASMBOY_GAME_RAM_BANKS_LOCATION: libWorker.wasmInstance.exports.gameRamBanksLocation.valueOf(),
-              WASMBOY_INTERNAL_STATE_SIZE: libWorker.wasmInstance.exports.wasmBoyInternalStateSize.valueOf(),
-              WASMBOY_INTERNAL_STATE_LOCATION: libWorker.wasmInstance.exports.wasmBoyInternalStateLocation.valueOf(),
-              WASMBOY_INTERNAL_MEMORY_SIZE: libWorker.wasmInstance.exports.gameBoyInternalMemorySize.valueOf(),
-              WASMBOY_INTERNAL_MEMORY_LOCATION: libWorker.wasmInstance.exports.gameBoyInternalMemoryLocation.valueOf(),
-              WASMBOY_PALETTE_MEMORY_SIZE: libWorker.wasmInstance.exports.gameboyColorPaletteSize.valueOf(),
-              WASMBOY_PALETTE_MEMORY_LOCATION: libWorker.wasmInstance.exports.gameboyColorPaletteLocation.valueOf()
-            },
-            eventData.messageId
-          )
-        );
-        return;
-      }
-      case WORKER_MESSAGE_TYPE.LOAD_ROM: {
-        // Load the game data into actual memory
-        libWorker.wasmByteMemory.set(new Uint8Array(eventData.message.ROM), libWorker.WASMBOY_GAME_BYTES_LOCATION);
-
-        // Forward to our lib worker
-        const wasmByteMemory = libWorker.wasmByteMemory.slice(0);
-        libWorker.memoryWorkerPort.postMessage(
-          getSmartWorkerMessage(
-            {
-              type: WORKER_MESSAGE_TYPE.LOAD_ROM_DONE,
-              wasmByteMemory: wasmByteMemory.buffer
-            },
-            eventData.messageId
-          ),
-          [wasmByteMemory.buffer]
-        );
-        return;
-      }
-    }
-  },
-
-  _controllerMessageHandler: event => {
-    // Handle our messages from the main thread
-    const eventData = getEventData(event);
-
-    switch (eventData.message.type) {
-      case WORKER_MESSAGE_TYPE.SET_JOYPAD_STATE: {
-        // Config will come in as an array, pass in values using apply
-        // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Function/apply
-        const setJoypadStateParamsAsArray = eventData.message.setJoypadStateParamsAsArray;
-        libWorker.wasmInstance.exports.setJoypadState.apply(libWorker, setJoypadStateParamsAsArray);
-        return;
-      }
-    }
-  },
-
-  _audioMessageHandler: event => {
-    // Handle our messages from the main thread
-    const eventData = getEventData(event);
-
-    switch (eventData.message.type) {
-      case WORKER_MESSAGE_TYPE.GET_CONSTANTS: {
-        libWorker.WASMBOY_SOUND_OUTPUT_LOCATION = libWorker.wasmInstance.exports.soundOutputLocation.valueOf();
-        // Forward to our lib worker
-        libWorker.audioWorkerPort.postMessage(
-          getSmartWorkerMessage(
-            {
-              type: WORKER_MESSAGE_TYPE.GET_CONSTANTS_DONE,
-              WASMBOY_SOUND_OUTPUT_LOCATION: libWorker.wasmInstance.exports.soundOutputLocation.valueOf()
-            },
-            eventData.messageId
-          )
-        );
-        return;
-      }
-
-      case WORKER_MESSAGE_TYPE.AUDIO_LATENCY: {
-        libWorker.currentAudioLatencyInSeconds = eventData.message.latency;
-        return;
-      }
-    }
   }
 };
